src/main/java/com/wechat/impl/ProductServiceImpl.java
package com.wechat.impl;

import com.wechat.dataobject.ProductInfo;
import com.wechat.repository.ProductInfoRepository;
import com.wechat.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;
@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    ProductInfoRepository productInfoRepository;
    @Override
    public ProductInfo findById(String ProductId) {
        return productInfoRepository.findById(ProductId).get();
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return productInfoRepository.findByProductStatus(0);
    }

    @Override
    public Page<ProductInfo> findall(Pageable pageable) {
        return productInfoRepository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return productInfoRepository.save(productInfo);
    }
}
